requirements.txt
pygments=2.18.0
pillow=10.3.0